backend/requirements.txt
asgiref==3.8.1
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.61.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.6.0
gunicorn==22.0.0
idna==3.7
oauthlib==3.2.2
packaging==24.1
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.13.3
virtualenv-clone==0.5.7
